+++ common/lib/CMakeLists.txt
@@ -39,7 +39,7 @@ endif()
 
 if(CMAKE_SYSTEM_NAME STREQUAL Linux)
     target_sources(villas-common PRIVATE
-        kernel/pci.cpp
+        kernel/devices/pci_device.cpp
         kernel/vfio_device.cpp
         kernel/vfio_group.cpp
         kernel/vfio_container.cpp
