+++ fpga/lib/pcie_card.cpp
@@ -14,7 +14,7 @@
 #include <villas/fpga/node.hpp>
 #include <villas/fpga/pcie_card.hpp>
 #include <villas/kernel/kernel.hpp>
-#include <villas/kernel/pci.hpp>
+#include <villas/kernel/devices/pci_device.hpp>
 #include <villas/kernel/vfio_container.hpp>
 #include <villas/memory.hpp>
